refactored docker_api
