+++ services/director-v2/src/simcore_service_director_v2/modules/dynamic_sidecar/docker_api/__init__.py
@@ -1,0 +1,47 @@
+from ._core import (
+    are_all_services_present,
+    constrain_service_to_node,
+    create_network,
+    create_service_and_get_id,
+    get_dynamic_sidecar_state,
+    get_dynamic_sidecars_to_observe,
+    get_or_create_networks_ids,
+    get_projects_networks_containers,
+    get_service_placement,
+    get_swarm_network,
+    inspect_service,
+    is_dynamic_service_running,
+    is_dynamic_sidecar_stack_missing,
+    list_dynamic_sidecar_services,
+    remove_dynamic_sidecar_network,
+    remove_dynamic_sidecar_stack,
+    try_to_remove_network,
+    update_scheduler_data_label,
+)
+from ._volume_removal import (
+    remove_pending_volume_removal_services,
+    remove_volumes_from_node,
+)
+
+__all__: tuple[str, ...] = (
+    "are_all_services_present",
+    "constrain_service_to_node",
+    "create_network",
+    "create_service_and_get_id",
+    "get_dynamic_sidecar_state",
+    "get_dynamic_sidecars_to_observe",
+    "get_or_create_networks_ids",
+    "get_projects_networks_containers",
+    "get_service_placement",
+    "get_swarm_network",
+    "inspect_service",
+    "is_dynamic_service_running",
+    "is_dynamic_sidecar_stack_missing",
+    "list_dynamic_sidecar_services",
+    "remove_dynamic_sidecar_network",
+    "remove_dynamic_sidecar_stack",
+    "remove_pending_volume_removal_services",
+    "remove_volumes_from_node",
+    "try_to_remove_network",
+    "update_scheduler_data_label",
+)
